requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1

# Logging e monitoramento
structlog==23.2.0
python-json-logger==2.0.7

# Cache Redis
redis==4.6.0
aioredis==2.0.1
fastapi-cache2[redis]==0.2.1

# Background tasks
celery==5.3.4
kombu==5.3.4

# Email (para notificações)
fastapi-mail==1.4.1
jinja2==3.1.2

# Utilitários adicionais
python-dateutil==2.8.2
pytz==2023.3